+++ mod/CHANGES.md
@@ -100,8 +100,10 @@ While there are thousands of rooms in the game, many players have already seen t
 * Identified pills (up to 7) will be shown when the player presses Tab.
 * Diagonal knife throws have a 3-frame window instead of a 1-frame window.
 * Duplicate rooms will no longer appear on the same run. (Basement 1 is exempt.)
+* Cod Worms are replaced with Para-Bites.
 * Wizoobs, Red Ghosts, and Lil' Haunts no longer have invulnerability frames after spawning.
 * Mom's Hands, Mom's Dead Hands, Wizoobs, and Red Ghosts have faster attack patterns.
+* Death will no longer perform his slow attack.
 * Betrayal no longer pauses the game and plays an animation before charming enemies.
 * The disruptive teleport that occurs when entering a room with Gurdy, Mom, Mom's Heart, or It Lives! no longer occurs.
 * The pickup delay on reloaded pedestal items is decreased from 18 frames to 15 frames.